requirements.txt
google-genai>=0.1.2
python-dotenv>=1.0.0
